限制演示账号不可增删改的操作

+++ src/main/java/com/jp/service/impl/UserServiceImpl.java
@@ -3222,7 +3222,12 @@ public class UserServiceImpl implements UserService {
 			return res;
 		}
 		String phone = entity.getPhone();
-
+		if ("18647740001".equals(phone)) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("不可编辑！");
+			res = new JsonResponse(result);
+			return res;
+		}
 		UserQuery userExample = new UserQuery();
 		userExample.or().andPhoneEqualTo(phone);
 		// userExample.setOrderByClause("createtime desc");
@@ -3298,6 +3303,12 @@ public class UserServiceImpl implements UserService {
 		if (entity.getPhone() == null || "".equals(entity.getPhone())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("请填写手机号");
+			res = new JsonResponse(result);
+			return res;
+		}
+		if ("18647740001".equals(entity.getPhone())) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("不可编辑！");
 			res = new JsonResponse(result);
 			return res;
 		}
@@ -3499,6 +3510,10 @@ public class UserServiceImpl implements UserService {
 	public JsonResponse bindingWithThirdParty(User entity, String smscode, Integer loginstatus) {
 		Result result = null;
 		JsonResponse res = null;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		if (StringUtils.isBlank(entity.getOpenid())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("参数openid为空");
@@ -3586,12 +3601,9 @@ public class UserServiceImpl implements UserService {
 	public JsonResponse relieveWithThirdParty(User entity) {
 		Result result = null;
 		JsonResponse res = null;
-		String userid = WebUtil.getHeaderInfo(ConstantUtils.HEADER_USERID);
-		if (userid == null || "".equals(userid)) {
-			result = new Result(MsgConstants.RESUL_FAIL);
-			result.setMsg("请求失败，header的userid为空！");
-			res = new JsonResponse(result);
-			return res;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
 		}
 		if (entity.getOpenid() == null || "".equals(entity.getOpenid())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
@@ -3599,7 +3611,7 @@ public class UserServiceImpl implements UserService {
 			res = new JsonResponse(result);
 			return res;
 		}
-		User user = userDao.selectByPrimaryKey(userid);
+		User user = userDao.selectByPrimaryKey(demoUser.getData().toString());
 		List<User> users = userDao.selectByPhone(user.getPhone());
 		if (users.size() > 0) {
 			for (User u : users) {
@@ -3862,6 +3874,12 @@ public class UserServiceImpl implements UserService {
 		if ("".equals(entity.getPhone()) || entity.getPhone() == null) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("没有获取到手机号！");
+			res = new JsonResponse(result);
+			return res;
+		}
+		if ("18647740001".equals(entity.getPhone())) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("不可编辑！");
 			res = new JsonResponse(result);
 			return res;
 		}
@@ -4280,6 +4298,10 @@ public class UserServiceImpl implements UserService {
 	public JsonResponse changeImgurl(User entity) {
 		Result result = null;
 		JsonResponse res = null;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		if ("".equals(entity.getImgurl()) || entity.getImgurl() == null) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("用户imgurl不存在！");
@@ -4310,6 +4332,10 @@ public class UserServiceImpl implements UserService {
 		Result result = null;
 		JsonResponse res = null;
 		int status = 0;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		if (entity.getUserid() == null || "".equals(entity.getUserid())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("用户ID不存在！");
@@ -4792,6 +4818,10 @@ public class UserServiceImpl implements UserService {
 		Result result = null;
 		JsonResponse res = null;
 		int icount = 0;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		if (userChildInfo.getPid() == null || "".equals(userChildInfo.getPid())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("缺少父亲ID信息！");
@@ -5024,6 +5054,10 @@ public class UserServiceImpl implements UserService {
 			res = new JsonResponse(result);
 			return res;
 		}
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		UserworkexpQuery userWorkexpExample = new UserworkexpQuery();
 		userWorkexpExample.or().andWorkidEqualTo(entity.getUpdateid());
 		int status = userworkDao.deleteByExample(userWorkexpExample);
@@ -5043,6 +5077,10 @@ public class UserServiceImpl implements UserService {
 	public JsonResponse deleteUserEduExp(User entity) {
 		Result result = null;
 		JsonResponse res = null;
+		JsonResponse demoUser = checkDemoUser();
+		if (demoUser.getCode() == 1) {
+			return demoUser;
+		}
 		if ("".equals(entity.getUpdateid()) || entity.getUpdateid() == null) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("教育经历id不存在！");
@@ -5085,6 +5123,12 @@ public class UserServiceImpl implements UserService {
 			res = new JsonResponse(result);
 			return res;
 		}
+		if ("18647740001".equals(entity.getPhone())) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("不可编辑！");
+			res = new JsonResponse(result);
+			return res;
+		}
 		if (StringUtils.isBlank(entity.getUsername())) {
 			result = new Result(MsgConstants.RESUL_FAIL);
 			result.setMsg("用户姓名不能为空，请重设后再试");
@@ -5116,7 +5160,7 @@ public class UserServiceImpl implements UserService {
 			return res;
 		}
 		//	List<User> byPhone = userDao.selectByPhone(entity.getPhone());
-		int status;
+		int status = 0;
 		try {
 			List<User> byPhoneAndStatus = userDao.selectByPhoneInStatus(entity.getPhone());
 			if (byPhoneAndStatus.size() >= 2) {
@@ -5131,13 +5175,12 @@ public class UserServiceImpl implements UserService {
 					return res;
 				}
 			}
-			List<User> byPhoneToStatus = userDao.selectByPhoneToStatus(entity.getPhone(), entity.getFamilyid());
-			if (byPhoneToStatus.size() >= 1) {
+			Integer count = userDao.selectByPhoneToStatus(entity.getPhone(), entity.getFamilyid());
+			if (count >= 1) {
 				result = new Result(MsgConstants.FAMILYID_REPULSE);
 				res = new JsonResponse(result);
 				return res;
 			}
-			status = 0;
 			// 创建用户
 			String userid = UUIDUtils.getUUID();
 			entity.setUserid(userid);
@@ -5259,5 +5302,30 @@ public class UserServiceImpl implements UserService {
 			onlineUsers.add(onLineUser);
 		}
 		return singleCorpLoginOrMultiCorpGetList(dbuserList, entity.getSessionid());
+	}
+
+	@Override
+	public JsonResponse checkDemoUser() {
+		Result result = null;
+		JsonResponse res = null;
+		//当前登录人 userid
+		String userid = WebUtil.getHeaderInfo(ConstantUtils.HEADER_USERID);
+		if (StringUtils.isBlank(userid)) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("用户非法！");
+			res = new JsonResponse(result);
+			return res;
+		}
+		String phone = userDao.selectUserPhone(userid);
+		if ("18647740001".equals(phone)) {
+			result = new Result(MsgConstants.RESUL_FAIL);
+			result.setMsg("不可编辑！");
+			res = new JsonResponse(result);
+			return res;
+		}
+		result = new Result(MsgConstants.RESUL_SUCCESS);
+		res = new JsonResponse(result);
+		res.setData(userid);
+		return res;
 	}
 }
